requirements.txt
langchain>=0.1.0
langchain-core>=0.1.0
langchain-community>=0.1.0
langgraph>=0.0.1
google-generativeai>=0.3.2
chroma-core>=0.4.0
chromdb>=0.4.0
python-dotenv>=1.0.0
fastapi>=0.109.0
uvicorn>=0.27.0
pydantic>=2.0.0
typings-core>=2.6.0